Added a new theme called 'ControlPanel'
It looks good, it's well rounded, highly functional. It properly indicates the current python version when python files are present, the venv name while active, it also contains a nodejs, npm, gcp, and git segments that are dynamically rendered under the appropriate conditions. It dynamically resizes, and everything makes sense.